Remove unnecessary codable conformances
