+++ Sources/LiveViewNative/Utils/Visibility.swift
@@ -14,7 +14,7 @@ import LiveViewNativeCore
 /// * `automatic`
 /// * `visible`
 /// * `hidden`
-extension Visibility: Decodable, AttributeDecodable {
+extension Visibility: AttributeDecodable {
     public init(from attribute: LiveViewNativeCore.Attribute?) throws {
         guard let attributeValue = attribute?.value else { throw AttributeDecodingError.missingAttribute(Self.self) }
         guard let value = Self(from: attributeValue) else { throw AttributeDecodingError.badValue(Self.self) }
@@ -28,12 +28,5 @@ extension Visibility: Decodable, AttributeDecodable {
         case "automatic": self = .automatic
         default: return nil
         }
-    }
-    
-    public init(from decoder: Decoder) throws {
-        let container = try decoder.singleValueContainer()
-        guard let value = Self(from: try container.decode(String.self))
-        else { throw DecodingError.dataCorrupted(.init(codingPath: container.codingPath, debugDescription: "expected valid value for Visibility")) }
-        self = value
     }
 }
